+++ HISTORY.rst
@@ -5,8 +5,9 @@ Release History
 
 0.1.49
 ++++++
-* Add Command Coverage Report  (#323)
+* Add Command Coverage Report for CLI modules (#323)
 * Add Linter rule missing_command_coverage and missing_parameter_coverage  (#323)
+* Add Command Coverage Report for CLI extensions (#383)
 
 0.1.48
 ++++++